add edit issue page
